// centroemprendimiento/src/app/interfaces/asistentetecnico.ts
export interface AsistenteTecnico {
        id_asistente_tecnico: number;
        fecha_registro_asistente_tecnico: string;
        estado_asistente_tecnico: string;
        id_edicion: number;
        edicion: string;
        fecha_inicio: string;
        fecha_fin: string;
        id_usuario_registro: number;
        id_usuario_modifica: number;
        fecha_modificacion_asistente_tecnico: string;
        id_persona: number;
        nombre: string;
        apellido: string;
        fecha_nacimiento: string;
        id_genero: number;
        genero: string;
        id_ciudad: number;
        ciudad: string;
        email: string;
        telefono: string;
        id_situacion_laboral: number;
        situacion_laboral: string;
        tipo_identificacion: string;
        identificacion: string;
        id_nivel_academico: number;
        nivel_academico: string;
        id_usuario: number;
        direccion: string;
        id_ciudad_domicilio: number;
        ciudad_domicilio: string;
        fecha_registro_persona: string;
        fecha_modificacion_persona: string;
        telefono_fijo: string;
        cv: string;
        estado_persona: string;
        uso_datos: string;
}